requirements.txt
einops
matplotlib
numpy
pandas
Pillow
scikit_learn
scipy
setuptools
torch
torchvision
tqdm